+++ api/src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 
-import { AuthModule } from '@Module/auth/auth.module';
-import { ProfileModule } from '@Module/profile/profile.module';
-import { ServiceModule } from '@Module/service/service.module';
-import { UserModule } from '@Module/user/user.module';
+import { AuthModule } from '@Src/modules/auth/auth.module';
+import { ProfileModule } from '@Src/modules/profile/profile.module';
+import { ServiceModule } from '@Src/modules/service/service.module';
+import { UserModule } from '@Src/modules/user/user.module';
 
 
 @Module({
