+++ api/src/modules/auth/interface/protected-request.interface.ts
