+++ api/src/modules/auth/session.service.ts
