+++ api/src/modules/profile/master-profile.service.ts
@@ -1,4 +1,4 @@
-import { MasterWeeklyScheduleService } from '@Module/profile/weekly-schedule.service';
+import { MasterWeeklyScheduleService } from '@Src/modules/profile/weekly-schedule.service';
 import { forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import Prisma from '@prisma/client';
