+++ api/src/modules/user/user-picture.service.ts
